+++ Src/Penneo/(Model)/Signer.cs
@@ -30,7 +30,7 @@ namespace Penneo
         public string ValidatedName { get; set; }
         public string SocialSecurityNumber { get; set; }
         /// <summary>
-        /// See https://app.penneo.com/api/v1/signers/ssn-types for available types
+        /// See https://app.penneo.com/api/v3/signers/ssn-types for available types
         /// </summary>
         public string SsnType = "legacy";
         public string OnBehalfOf { get; set; }